bp_genbank_ref_extractor: skip genes of any type other than 'pseudo' or 'protein-coding'.
* the current code was checking the gene type, and treat it as pseudo gene if type was
'pseudo', and as protein coding otherwise. However, this approach ignores some less
common cases such as long non-coding RNA (e.g., H19 with UID 14955), and even regions
of the genome (with type of 'other') which sometimes appear annotated as genes (e.g.,
Hist3 with UID 15268). There may be other gene types that which we are not aware of,
so we limit ourselves to coding and pseudo genes.